Move all tests ina rticle except meta tests
